Integrated buildpack
Previously, the buildpack was split across multiple repositories.  As
this was a premature optimization, the code has been combined into a
single repository for now.  This change also adds a default behavior
of executing a Java application based on the Main-Class value in its
manifest.  This is not intended to be the behavior long-term and is
only there to facilitate current testing.

[#49321577]