Externalize JRE configuration to configuration file
Previously, the JRE configuration (URIs, etc.) were embedded in a Ruby
class.  This was not easy to extend for buildpack developers.  This
change moves that configuration into a YAML file located in
config/jres.yml and introduces a new data structure.  This data
structure specifies a URI that points at the root of a repository of
JREs.  The code now expects that there is an index.yml file located at
this root, and reads this in order to know what files to download.

In addition to externalizing this configuration, this change
introduced a wildcard syntax for JRE versions.  The syntax for JRE
versions, including the wildcards, is much better policed now.

To make these changes, much of the underlying infrastructure was
refactored into much smaller modules.

[#50171283]