Choose OpenJDK 1.7.0_+ if no JRE specified
Previously, if no JRE was specified, the absolutely latest version of
OpenJDK was selected.  This resulted in a milestone version of 1.8
being selected and was not what users would have expected.  This
change allows the maintainer of config/jres.yml to explicitly specify
a version (with wildcards) that should act as the default version to
user for a JRE vendor.

[#49322113]