Incorrect working directory assumption
Previously, the code in the JreSelector that loaded config/jres.yml
assumed that the working directory of the process was the root of the
buildpack.  This assumption is incorrect when running in Cloud
Foundry.  This change adjusts references to the configuration file to
be relative to the file requesting it.

[#50980505]