Tomcat container
Previously, web applications had no container to run in.  This change
adds a Tomcat container to run web applications in.  The container
detects the presence of a WEB-INF directory in the application,
downloads and configures Tomcat during the compile phase, and returns
a Tomcat starting command in the release phase.

[#49694799]