Relative instead of absolute paths
Previously, the buildpack used absolute paths everywhere, including in
the command that was run for the droplet.  This was incorrect as the
absolute path during staging is different than the absolute path at
runtime.  This changes the command to always use relative paths while
preserving absolute paths for the mutation work done as part of
compile.

[#49694799][#1]