Configure buildpack support JAR in Tomcat
Download the JAR from its repository. Add the JAR to Tomcat's lib
directory and define a listener in server.xml.

The current version of the support JAR provides a listener which
shuts down Tomcat if the web application fails to start. This
improves the diagnostics when pushing such an application to Cloud
Foundry as Cloud Foundry checks the health of Tomcat rather than of
the application.

[#49091441]