Remove system.properties
Previously, the buildpack could be configured by adding a
system.properties file into an application before it was pushed.  This
conflated the actions of pushing and application and configuring the
environment that the application ran in.  This change removes most of
the use of system.properties from the system, but does not quite
remove it all.  The JRE memory subsystem still has references to
system.properties (although those code-paths will never be called) and
those will be removed in another story.

In addition to the code changes around system.properties, the
structure of the documentation has been changed.  The documentation is
now in multiple Markdown files located in the docs/ directory.  These
files are linked to from a table of contents embedded in README.md.
This means that a user will see the documentation links on first
arrival at the buildpack and will be sent to more detailed information
that is versioned with the buildpack code itself.

[#52105249]