Remove system.properties support
Previously the memory subsystem relied on system.properties support.
Now that it has been removed, this area must change.  This change
restructures the code slightly, and removes the need to declare
mappings for both types of JRES.  It also ensures that knowledge of
valid configuration parameters is handled via polymorphism.

[#52129161]