Improved OOM kill support
Passing

-XX:OnOutOfMemoryError='kill -9 %p'

to Java via the standard Play application start script cannot be done.

Bash ruins whatever combination of escapes are passed in and typically
java fails saying that the option -9 is invalid. See
http://stackoverflow.com/questions/5792049/xxonoutofmemoryerror-
kill-9-p-problem for general, if inconclusive, discussion of the
general problem.

To work around this problem, introduce a killjava script which
OpenJdk.compile places in the buildpack diagnostics directory.
Specify this script on -XX:OnOutOfMemoryError.

Also, improve test coverage.

[#51899933]