[#52324579]Add migration guide
The migration guide documents how to migrate from the previous Java
buildpack.

The migration notes include intentional migration impacts and some
features which are still required to bring the new buildpack up to
parity with the old buildpack. The latter are flagged as temporary
restrictions.

Note in the OpenJDK documentation that $MEMORY_LIMIT is not yet
passed to the buildpack.