Update diagnostic logging
This change includes a number of improvements to the diagnostic
logging.  The first is that the logs are no put into the
.buildpack_diagnostics hidden directory.  This is a visual indicator
that the content is contributed by the buildpack and not the user. The
second improvement is that logging is now formatted such that each log
message has a time and title, followed by an indented payload if there
is a payload.  Finally, the logging code has been refactored out into
a utility class.

All of this is simply a work around until a reasonable logging
framework is provided by Cloud Foundry core.

[#52706275]