Add Spring Auto-reconfiguration JAR to Spring apps
The change adds a new framework that detects Spring applications and
adds the Spring Auto-reconfiguration JAR to them.  This change
required the introduction of a shared '.lib' directory that frameworks
can contribute JARs to.  The containers are then responsible for
ensuring that the JARs in this shared directory are put on the
classpath at runtime.  As part of this change, all of the existing
containers (main, play, and tomcat) have been update to use JARs in
this directory.

[#49695345]