Modify web.xml for auto-reconfiguration
Previously, the auto-reconfiguration JAR was simply copied into the
additional lib directory for any application that had Spring in it.
This did not actually enable any of the auto-reconfiguration
functionality.  This change causes the auto-reconfiguration framework
to modify a /WEB-INF/web.xml file if it exists in the application.
The modifications made by the framework are detailed in the
documentation for it.

[#49695353]