Class-Path manifest header
The change updates the java-main container to use the Class-Path
manifest header.  The container examines the application for that
header and if it exists, places all values from it on the classpath
when running the application.  To facilitate this, the properties
support was expanded to handle JAR manifest line-continuation
semantics.

[#52899629]