Play version
Previously, the detect output of the Play container was simply 'play'
without a version of the Play Framework being used.  As the detect
output is intended to be used to determine when applications a
susceptible to security exploits, it was important to output that
version.  This change attempts to determine the version by looking at
the Play JAR and extracting it's version number from it.

[#52855559]