Add auto reconfiguration to Play applications
Previously, auto reconfiguration was limited to web applications only.
This was not by design, but was the case in practice.  In the auto
reconfiguration support, Play applications can be reconfigured
(although in a more limited way) and this support was not used
properly.  This change splits the auto-reconfiguration support into a
'Spring' side and a 'Play' side.  The supported JAR is the same
regardless of which kind of reconfiguration you are getting, but the
detection and modification behaviors are quite different.