Play support
This change updates the existing play support to actually work.  The
first change was to cleanup the tests such that they never write into
the test source tree.  The second change was to attach additional
libraries into the classpath differently for the dist versus staged
payloads; added to the classpath variable for dist, symbolically
linked into staged for staged.  The third change was to replace the
default Netty server with our Bootstrap implementation.

[#49696097]