Ensure missing cached file behavior
Previously, the DownloadCache did the right thing (re-download) a file
when it had been moved out of the cached location as part of block
execution.  However, it was not clear from the test code that this was
its intended behavior.  This change adds a test that explicitly tests
this behavior.

[#52124997]