Add New Relic support
This change adds support for bound New Relic services.  The framework
notices a bound service in VCAP_SERVICES.  During the compile phase,
the New Relic agent JAR is download and a default New Relic
configuration file is copied into place.  During the release phase, a
number of JAVA_OPTS are configured such that the New Relic service is
used during the application's runtime.

[#53027669]