Test environments
This change adds support for test environments.  The targeted test
environments are lucid64, precise64, and mountainlion64.  The support
comes in the form of the ability to resolve the platform and
architecture of the current host when resolving a repository URI.  The
resolution allows the repository to contain wildcards of '{platform}'
and '{architecture}' which are then substituted with the appropriate
value.

The other half of this support is the modification of the existing
repositories.  Since our Java dependencies do not currently depend on
the host that they are running on, their repository URIs have been
changed to remove the platform and architecture directories.  The
OpenJDK dependency is platform and architecture specific and as such,
it's URI has been modified to include the placeholders.

The final step is the update of all of the documentation to accurately
reflect the new repository structures.

[#52860123]