Zero-touch support for Spring Insight
Spring Insight is a monitoring solution provided Pivotal[1].  It monitors
Spring web applications for performance and faults.  This support is
implemented as a framework that looks for a service called 'insight' and acts
when it is found.  The framework downloads the appropriate Insight agent from
the Insight dashboard and configures it so that it connects to the bound
service properly.  Configuration of Insight is not accomplished through the
buildpack, but rather through the Insight dashboard.  This service only works
for applications deploying into the Tomcat container and modifies the Tomcat
container with *TEMPORARY* support for behaviors needed by the Spring Insight
agent.  These changes to the Tomcat container are not likely to last long and
other components should not depend on or emulate them.

[1]: http://gopivotal.com/pivotal-products/apps/pivotal-tc-server

[52064679] [52065001] [52065073]