Fix a path relativity issue in Spring Insight
A previous commit [1] added support for the Spring Insight application
performance monitoring tool to the Tomcat container. This commit
contained an error where the installed location of Insight was
referenced with an absolute path instead of a relative path. This
causes applications bound to the Insight service to fail to start as
the path to Insight passed to the application at startup will be
incorrect.

[1] e0e90410aaed1b87d791ad031f1577409f90614a - Zero-touch support for
Spring Insight