Remove yard from the default build
Previously yard would be run for the default build (run during CI).
However the artifacts created are never persisted so it was a waste a
non-trivial of time.  The change keeps the API enforcement task in the
build, but removes the doc generation task.  To create the docs,
simply run `bundle exec rake yard`.