[#59817096]Support Errno::ENETUNREACH
The internet availability check was failing because Errno::ENETUNREACH
was not being handled.

This fix adds Errno::ENETUNREACH and a number of other network-related
exceptions to the set being handled. When these errors are caught on
the first download attempt in a buildpack invocation (detect, compile,
or release), the internet is deemed to be inaccessible and a look-
aside is performed in the (currently read-only, but pre-populated)
buildpack cache.

See https://www.pivotaltracker.com/s/projects/803319/stories/58941750
for background.

[#59817096]