Default repository root
Previously, all of the repository_root declarations in the component
configurations used the common
http://download.pivotal.io.s3.amazonaws.com domain for their
repositories.  Since any user wanting to move the repositories to a
different location (e.g. replication behind a firewall) would have to
change many files, this change abstracts the repository root
declaration.  It does this by introducing a {default.repository.root}
substitution variable for repository_roots.  This placeholder is
substituted with a value configured in config/repository.yml.  The use
of this placeholder is not required.

[#57585912]