JARs added to additional libraries in compile
Previously, some of the frameworks only added their JARs to the
additional_jars collection as part of release.  Since some components
symbolically link the additional JARs (as part of compile), these JARs
were not showing up on the classpath.  This change ensures that the
JARs are added  as part of compile.

[#61738342]