Find JDBC driver JARs in subdirectories
Previously, the JDBC driver frameworks only looked into the root of an
application to find the JDBC driver JARs that would cause them to
avoid.  This was an error as many times those JAR files would be in a
nested subdirectory (e.g. WEB-INF/lib).  This change fixes this bug
and causes the JDBC driver JARs to be detected anywhere in the
application's filesystem.

[#63555022]