Allow Spring Insight to support multiple app instances
Spring Insight is a monitoring solution provided Pivotal[1].  It monitors
Spring web applications for performance and faults.  This support is
implemented as a framework that looks for a service called 'insight' and acts
when it is found.  The framework downloads the appropriate Insight agent from
the Insight dashboard and configures it so that it connects to the bound
service properly.  Each instance of an application is configured with the
same name and Spring Insight can not tell the difference between instances.
This commit fixes this by removing the application naming property supplied
to Insight at startup and adding a new plugin to the Insight agent that names
application activity from different instances based on instance name set in
the Cloudfoundry enviroment.

[1]: http://gopivotal.com/pivotal-products/apps/pivotal-tc-server

[62015878]