OracleJRE support
This change adds Oracle JRE support to the buildpack.  This support is
identical to the OpenJDK support an as such, uses an identical code-
path.  It does differ in that there are no pre-populated repositories
holding the binaries for the Oracle JRE.  This is due to licensing
restrictions and legal requirements.  By default the Oracle JRE is
disabled, but the documentation includes instructions on how to create
a repository and enable the Oracle JRE support.

[#53593605]