Use Yard macros in API docs
Previously there was a hole in our API documentation that overridden
methods did not inherit their parent's documentation.  There is an
issue open with the Yard team to fix this bug, but until then the
decision was made to use Yard macros to duplicate the documentation.
This change adds the Yard macros and cleans up other miscellaneous
documentation errors.

[#59111410]